refactor animation module
